frontend/backend/requirements.txt
flask==3.0.3
opencv-python==4.10.0.84
requests==2.32.3
google-generativeai
pillow
